// Sistema_Basico/src/pages/horarios/components/type.ts
export interface ModalConfirmacionProps {
  abierto: boolean;
  onConfirmar: () => void;
  onCancelar: () => void;
  titulo?: string;
  mensaje?: string;
  tipo?: "eliminar" | "cancelar" | "reprogramar";
}